rename Components to pascal case
